Move entity classes
