+++ src/main/java/pay/jh/me/moneysprinkling/entity/MoneySprinkle.java
@@ -1,12 +1,6 @@
-package pay.jh.me.moneysprinkling.model;
+package pay.jh.me.moneysprinkling.entity;
 
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
+import lombok.*;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -15,10 +9,9 @@ import javax.persistence.Table;
 
 @Getter
 @Setter
-@EqualsAndHashCode
-@ToString
 @NoArgsConstructor
-@RequiredArgsConstructor
+@AllArgsConstructor
+@Builder
 @Entity
 @Table(name = "money_sprinkle")
 public class MoneySprinkle {
